+++ src/main/java/client/RaceResultObserver.java
@@ -5,5 +5,4 @@ import protocol.RaceResult;
 public interface RaceResultObserver {
 
   void receivedRaceResult(RaceResult result);
-
 }