+++ src/main/java/protocol/RaceResult.java
@@ -4,14 +4,10 @@ import java.util.List;
 
 public class RaceResult {
 
-  /**
-   * Duration of race.
-   */
+  /** Duration of race. */
   public final long duration;
 
-  /**
-   * List of players ordered from best to worst.
-   */
+  /** List of players ordered from best to worst. */
   public List<UserResult> classification;
 
   public RaceResult(long duration, List<UserResult> classification) {